cloud-run/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Google Cloud
google-cloud-firestore==2.16.0
firebase-admin==6.5.0
google-generativeai==0.7.0

# Utilities
pydantic==2.7.0
pydantic-settings==2.2.0
python-dotenv==1.0.1
httpx==0.27.0

# Date/Time
python-dateutil==2.9.0
